feat: prevent the lock screen from requiring a password
It doesn't make sense to require a password when setting up the first user, so we disable that
requirement.

We thought of using something like DBus to inhibit the lock screen, but that could be problematic,
for example in a situation where the user leaves their laptop unattended during setup and instead of
sleeping it runs down the battery.