fix: disable session transition
Instead of transitioning automatically to the new user's account we will
simply log out, and allow them to enter their password at the login screen.
This will resolve the issue with the new user's wallet not being available
until after a reboot, until we have time to implement a proper solution
that allows us to retain the automatic session transition.

Resolves https://invent.kde.org/plasma/kiss/-/issues/26